refactor: 추천 조합, 좋아요 많은 조합 위젯 분리
